Refactoring to improve performance
